submitted/thread_21383/warning/fix-signed-unsigned-mismatch
Fixes warning like 'expression' : signed/unsigned mismatch

+++ src/OSD/OSD_MAllocHook.cxx
@@ -494,7 +494,7 @@ void OSD_MAllocHook::CollectBySize::Reset()
     myArray = new Numbers[MAX_ALLOC_SIZE];
   else
   {
-    for (int i = 0; i < MAX_ALLOC_SIZE; i++)
+    for (int i = 0; i < (int)MAX_ALLOC_SIZE; i++)
       myArray[i] = Numbers();
   }
   myTotalLeftSize = 0;
@@ -525,7 +525,7 @@ Standard_Boolean OSD_MAllocHook::CollectBySize::MakeReport(const char* theOutFil
            << std::setw(20) << "PeakSize " << std::endl;
 
   Standard_Size aTotAlloc = 0;
-  for (int i = 0; i < MAX_ALLOC_SIZE; i++)
+  for (int i = 0; i < (int)MAX_ALLOC_SIZE; i++)
   {
     if (myArray[i].nbAlloc > 0 || myArray[i].nbFree > 0)
     {
